Support chunking files for Vite applications
Fix #60

Added support for a chucked Vite application.

Ensure the index....js file is first in the list of js files. Sort the
remaining files alphabetically. Doing so allows a develop to control the
order by prefixing the chunk names to obtain a desired load order.

When handling the script_loader_tag filter, use <script> to load the
index.js file and <link> with rel="modulepreload" for the chunk files.